backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
drf-spectacular==0.27.2
whitenoise==6.8.2
gunicorn==23.0.0
